first working version

+++ app/src/main/java/edu/nav/hermes/math/algorithms/Answer.java
@@ -10,6 +10,6 @@ import java.util.List;
  * Answer of the algorithm
  */
 public class Answer {
-    public float cost = 0;
+    public double cost = 0;
     public List<IGeoPoint> path = new ArrayList<>();
 }
